backend/requirements.txt
Django
psycopg2-binary
python-logstash